requirement.txt
python-dotenv
flask
pymongo
OTXv2
humanize
python-magic
stix2
taxii2-client
medallion
yara-python